+++ docs/conf.py
@@ -10,7 +10,7 @@ project = 'Verisocks'
 copyright = '2023-2026, Jérémie Chabloz'
 author = 'Jérémie Chabloz'
 version = '1.7'
-release = '1.7.0-dev'
+release = '1.7.0'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration